requirements.txt
# Streamlit
streamlit>=1.30.0

# Google Gemini ADK and AI
google-genai
google-adk

# Pydantic for structured outputs
pydantic>=2.0.0

# Additional utilities
python-dotenv
